+++ plugins/map/constant.js
