+++ plugins/map/index.js
